Feature/Chatting Emoji

+++ network/message/src/main/java/com/seugi/network/message/response/message/MessageEmojiResponse.kt
@@ -5,5 +5,5 @@ import com.seugi.network.message.response.MessageTypeResponse
 data class MessageEmojiResponse(
     override val type: String,
     val emojiId: Int,
-    val userId: List<Int>,
+    val userId: List<Long>,
 ) : MessageTypeResponse()
